#22 이슈 010 - 코드리뷰반영
