+++ src/main/java/com/flab/tour/domain/user/service/UserService.java
@@ -11,7 +11,6 @@ import com.flab.tour.db.user.UserUpdateMapper;
 import com.flab.tour.domain.user.controller.model.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
 @Service
@@ -30,7 +29,7 @@ public class UserService extends BaseService {
         }
     }
 
-    public UserResponse signinByEmail(UserRegisterRequest svo) {
+    public UserResponse signupByEmail(UserRegisterRequest svo) {
         userRepository.findFirstByEmail(svo.getEmail())
                 .ifPresent(user -> {
                     throw new ApiException(UserErrorCode.EXIST_ID);
@@ -47,15 +46,14 @@ public class UserService extends BaseService {
         return ObjectConvertUtil.getInstance().copyVO(userEntity, User.class);
     }
 
-    @Transactional
-    public UserResponse updateProfile(UUID userId, UserUpdateRequest request) {
+    public UserResponse updateProfile(String userId, UserUpdateRequest request) {
         var userEntity = getUserWithThrow(userId);
         userUpdateMapper.updateUserFromRequest(request, userEntity);
         userRepository.save(userEntity);
         return new UserResponse("프로필 수정완료");
     }
 
-    public UUID getUserId(UserLoginRequest request) {
+    public String getUserId(UserLoginRequest request) {
         return userRepository.findFirstByEmailAndPasswordOrderByUserIdDesc(request.getEmail(), request.getPassword())
                 .orElseThrow(() -> new ApiException(UserErrorCode.USER_NOT_FOUND)).getUserId();
     }
